+++ src/lsp_data.rs
@@ -200,13 +200,13 @@ pub struct RequestMessage<T>
     where T: Debug + Serialize
 {
     jsonrpc: &'static str,
-    pub id: u64,
+    pub id: u32,
     pub method: String,
     pub params: T,
 }
 
 impl <T> RequestMessage<T> where T: Debug + Serialize {
-    pub fn new(id: u64, method: String, params: T) -> Self {
+    pub fn new(id: u32, method: String, params: T) -> Self {
         RequestMessage {
             jsonrpc: "2.0",
             id,
